# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key:
        description: Key is the plaintext key, set only in the response that creates
          it.
        type: string
      label:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  models.ChangePasswordResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      label:
        type: string
    type: object
  models.CreateAPIKeyResponse:
    properties:
      api_key:
        $ref: '#/definitions/models.APIKey'
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      must_change_password:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:5174
info:
  contact: {}
  description: REST API for sending and receiving SMS messages via a USB GSM modem.
  license:
    name: GPL-3.0
  title: SMS Gateway API
  version: "1.0"
paths:
  /api/v1/apikeys:
    get:
      description: Returns API keys for the authenticated user, newest first. Admins
        see all keys.
      parameters:
      - description: Maximum keys to return (max 500). Omit to return all.
        in: query
        name: limit
        type: integer
      - description: Keys to skip. Only applied together with limit.
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Total matching keys is returned in the X-Total-Count header
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Generates a new API key for the authenticated user with an optional
        label.
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
  /api/v1/apikeys/{id}:
    delete:
      description: Deactivates an API key by its unique identifier. Non-admins can
        only deactivate their own keys.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: API key deactivated'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate API key
      tags:
      - API Keys
  /api/v1/apikeys/{id}/delete:
    delete:
      description: Permanently deletes an API key by its unique identifier. Non-admins
        can only delete their own keys.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: API key deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - API Keys
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Allows an authenticated user to change their password by providing
        the current and new passwords. Every existing token is revoked; the response
        carries a new one for the caller.
      parameters:
      - description: Current and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChangePasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with username and password, returning a JWT
        token.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Revokes every token issued to the authenticated user, logging them
        out on all devices.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: logged out'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    description: API key for programmatic access
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix (e.g. "Bearer eyJhbG...")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"